+++ bin/data_generator.dart
@@ -1,7 +1,9 @@
+import 'dart:convert';
 import 'dart:io';
 import 'package:http/http.dart' as http;
 
 import 'config.dart';
+import 'micro_list_data.dart';
 
 Future<void> generateData(GenerationConfig config, bool dryRun) async {
   final currentDir = Directory.current.path;
@@ -13,8 +15,8 @@ Future<void> generateData(GenerationConfig config, bool dryRun) async {
   }
   outFile.createSync();
   var sink = outFile.openWrite();
-  sink.write("import 'types.microcms.g.dart';\n");
-  sink.write("import 'package:static_micro_cms/static_micro_cms.dart';\n\n");
+  sink.write("import 'dart:convert';\n");
+  sink.write("import 'types.microcms.g.dart';\n\n");
   final staticDataDeclarations = <String>[];
   for (var api in config.apis) {
     final privateName = "_\$${api.endpoint}Data";
@@ -22,7 +24,24 @@ Future<void> generateData(GenerationConfig config, bool dryRun) async {
       sink.write("const $privateName = r'';\n");
     } else {
       final res = await getMicroData(config, api.endpoint);
-      sink.write("const $privateName = r'$res';\n");
+      final String serialized;
+      if (api.type == ApiType.list) {
+        var parsed = MicroListData.fromString(res);
+        // get all contents
+        final contentList = parsed.contents;
+        var offset = parsed.limit;
+        while (offset < parsed.totalCount) {
+          final pageRes =
+              await getMicroData(config, api.endpoint, offset: offset);
+          parsed = MicroListData.fromString(pageRes);
+          contentList.addAll(parsed.contents);
+          offset += parsed.limit;
+        }
+        serialized = jsonEncode(contentList);
+      } else {
+        serialized = res;
+      }
+      sink.write("const $privateName = r'$serialized';\n");
     }
     final upperName = api.endpoint[0].toUpperCase() + api.endpoint.substring(1);
     final lowerName = api.endpoint[0].toLowerCase() + api.endpoint.substring(1);
@@ -31,7 +50,7 @@ Future<void> generateData(GenerationConfig config, bool dryRun) async {
           "static final ${upperName}MicroData ${lowerName}Data = ${upperName}MicroData.fromString($privateName);");
     } else if (api.type == ApiType.list) {
       staticDataDeclarations.add(
-          "static final MicroListData<${upperName}MicroData> ${lowerName}Data = MicroListData.fromString($privateName, (raw) => ${upperName}MicroData.fromString(raw));");
+          "static final List<${upperName}MicroData> ${lowerName}Data = (jsonDecode($privateName) as List).map((e) => ${upperName}MicroData.fromString(jsonEncode(e))).toList();");
     }
   }
   final dataStoreDeclaration = """
@@ -43,9 +62,13 @@ class MicroCMSDataStore {
   sink.close();
 }
 
-Future<String> getMicroData(GenerationConfig config, String endpoint) async {
+Future<String> getMicroData(GenerationConfig config, String endpoint,
+    {int? offset}) async {
   var url = config.baseUrl + "/$endpoint";
-  var res = await http
-      .get(Uri.parse(url), headers: {"X-MICROCMS-API-KEY": config.apiKey});
+  var uri = Uri.parse(url);
+  if (offset != null) {
+    uri = uri.replace(queryParameters: {"offset": "$offset"});
+  }
+  var res = await http.get(uri, headers: {"X-MICROCMS-API-KEY": config.apiKey});
   return res.body;
 }