+++ lib/src/base_objects.dart
@@ -1,5 +1,3 @@
-import 'dart:convert';
-
 abstract class MicroObjectData {
   final DateTime createdAt;
   final DateTime updatedAt;
@@ -14,25 +12,6 @@ abstract class MicroListItem extends MicroObjectData {
   const MicroListItem(this.id, DateTime createdAt, DateTime updatedAt,
       DateTime publishedAt, DateTime revisedAt)
       : super(createdAt, updatedAt, publishedAt, revisedAt);
-}
-
-class MicroListData<T extends MicroListItem> {
-  final List<T> contents;
-  final int totalCount;
-  final int offset;
-  final int limit;
-
-  const MicroListData(this.contents, this.totalCount, this.offset, this.limit);
-
-  static MicroListData<E> fromString<E extends MicroListItem>(
-      String rawData, E Function(String raw) contentGen) {
-    final m = jsonDecode(rawData);
-    return MicroListData(
-        (m["contents"] as List).map((e) => contentGen(jsonEncode(e))).toList(),
-        m["totalCount"],
-        m["offset"],
-        m["limit"]);
-  }
 }
 
 class MicroImageData {