+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.0
+
+- accept list apis that have more than 10 items
+
+Breaking:
+- change type definition of list apis to List<>
+- remove properties of list apis: totalCount, offset and limit
+
 ## 0.1.1
 
 - fix to use baseUrl